refactored apis methods
