+++ app/src/main/java/ru/itis/morefy/core/data/api/SpotifyAlbumsApi.kt
@@ -1,0 +1,39 @@
+package ru.itis.morefy.core.data.api
+
+import retrofit2.http.GET
+import retrofit2.http.Path
+import retrofit2.http.Query
+import ru.itis.morefy.core.data.response.album.*
+
+interface SpotifyAlbumsApi {
+    // albums - https://developer.spotify.com/documentation/web-api/reference/#/operations/get-an-album
+
+    @GET("albums/{id}")
+    fun getAlbumById(@Path("id") id: String, @Query("market") market: String) : AlbumResponse
+
+    @GET("albums")
+    fun getAlbumsByIds(@Query("ids") ids: String, @Query("market") market: String) : AlbumsResponse
+
+    @GET("albums/{ids}/tracks")
+    fun getTracksByAlbumId(@Path("id") ids: String, @Query("market") market: String) : AlbumTracksResponse
+
+    @GET("albums/{ids}/tracks")
+    fun getTracksByAlbumId(@Path("id") ids: String, @Query("limit") limit: Int, @Query("offset") offset: Int, @Query("market") market: String) : AlbumTracksResponse
+
+    @GET("me/albums")
+    fun getUserSavedAlbums(@Query("market") market: String) : UserSavedAlbumsResponse
+
+    @GET("me/albums")
+    fun getUserSavedAlbums(@Query("market") market: String, @Query("limit") limit: Int, @Query("offset") offset: Int) : UserSavedAlbumsResponse
+
+    @GET("browse/new-releases")
+    fun getNewReleasesAlbums(@Query("country") country: String) : NewReleasesAlbumsResponse
+
+    @GET("me/albums/contains")
+    fun checkIfAlbumsSaved(@Query("ids") albumsIds: String) : CheckSavedAlbumsResponse
+
+    @GET("browse/new-releases")
+    fun getNewReleasesAlbums(@Query("country") country: String, @Query("limit") limit: Int, @Query("offset") offset: Int) : NewReleasesAlbumsResponse
+
+
+}
