+++ app/src/main/java/ru/itis/morefy/core/data/request/CreatePlaylistRequestDto.kt
@@ -1,0 +1,8 @@
+package ru.itis.morefy.core.data.request
+
+data class CreatePlaylistRequestDto (
+  private val name: String,
+  private val public: Boolean,
+  private val collaborative: Boolean,
+  private val description: String,
+)
