+++ app/src/main/java/ru/itis/morefy/core/data/request/DeletePlaylistTracksDto.kt
@@ -1,0 +1,10 @@
+package ru.itis.morefy.core.data.request
+
+data class DeletePlaylistTracksDto (
+    private val tracks: List<TracksRequestDto>,
+    private val snapshot_id: String,
+)
+
+data class TracksRequestDto (
+    private val uri: String,
+)
