+++ app/src/main/java/ru/itis/morefy/core/data/response/CreatePlaylistResponse.kt
@@ -1,0 +1,5 @@
+package ru.itis.morefy.core.data.response
+
+class CreatePlaylistResponse {
+
+}
