+++ app/src/main/java/ru/itis/morefy/core/data/response/UpdateTracksResponse.kt
@@ -1,0 +1,5 @@
+package ru.itis.morefy.core.data.response
+
+class UpdateTracksResponse {
+// snapshot_id
+}
